finished implenting sounds, improved sound manager
finished implenting every sound possible with availble code and added an example song for testing. also improved sound manager to add a stopping function that stops a given sound and a stop looping function that lets a sound finish but stops it from looping if it was set to loop.