you now carry over what you have into level 2
refactored saving and loading to make it so you carry over what you have from level 1 to level 2, using the saving and loading system. i need to push the scene to make the level trigger have the player attached. it's also a prefab now as well as being a 2D item and not a cube